# docs/swagger/swagger.yaml
basePath: /api
definitions:
...
  models.SysRoleAddRequest:
    properties:
      description:
        type: string
      name:
        type: string
      parentId:
        type: integer
      sort:
        type: integer
      status:
        type: integer
    required:
    - name
    type: object
  models.SysRoleDataScopeUpdateRequest:
    properties:
      checkedDepts:
        type: string
      dataScope:
        type: integer
      id:
        type: integer
    required:
    - id
    type: object
  models.SysRoleMenuAssignRequest:
    properties:
      menuId:
        items:
          type: integer
        type: array
      roleId:
        type: integer
    required:
    - menuId
    - roleId
    type: object
  models.SysRoleUpdateRequest:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      parentId:
        type: integer
      sort:
        type: integer
      status:
        type: integer
    required:
    - id
    - name
    type: object
...
host: localhost:8080
info:
  contact:
    email: your-email@example.com
    name: API Support
    url: https://github.com/your-repo/gin-fast/issues
  description: 基于Gin框架的快速开发脚手架API文档
  license:
    name: MIT
    url: https://github.com/your-repo/gin-fast/blob/master/LICENSE
  termsOfService: https://github.com/your-repo/gin-fast
  title: Gin-Fast API
  version: "1.0"
paths:
...
  /sysRole/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定角色
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 角色删除成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 角色ID格式错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 删除角色
      tags:
      - 角色管理
    get:
      consumes:
      - application/json
      description: 根据角色ID获取角色详细信息
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回角色信息
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 角色ID格式错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 根据ID获取角色信息
      tags:
      - 角色管理
  /sysRole/add:
    post:
      consumes:
      - application/json
      description: 创建新角色
      parameters:
      - description: 角色信息
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/models.SysRoleAddRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 角色创建成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 新增角色
      tags:
      - 角色管理
  /sysRole/addRoleMenu:
    post:
      consumes:
      - application/json
      description: 为指定角色分配菜单权限
      parameters:
      - description: 角色菜单信息
        in: body
        name: roleMenu
        required: true
        schema:
          $ref: '#/definitions/models.SysRoleMenuAssignRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 分配角色菜单权限成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 为角色分配菜单权限
      tags:
      - 角色管理
  /sysRole/edit:
    put:
      consumes:
      - application/json
      description: 更新角色信息
      parameters:
      - description: 角色信息
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/models.SysRoleUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 角色更新成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 更新角色
      tags:
      - 角色管理
  /sysRole/getRoles:
    get:
      consumes:
      - application/json
      description: 获取所有角色列表（树形结构）
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回角色列表
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 获取角色列表
      tags:
      - 角色管理
  /sysRole/getUserPermission/{roleId}:
    get:
      consumes:
      - application/json
      description: 根据角色ID获取该角色拥有的菜单权限
      parameters:
      - description: 角色ID
        in: path
        name: roleId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回角色菜单权限
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 角色ID格式错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 根据角色ID获取菜单权限
      tags:
      - 角色管理
  /sysRole/list:
    get:
      consumes:
      - application/json
      description: 获取角色列表，支持分页和过滤
      parameters:
      - default: 1
        description: 页码
        in: query
        name: pageNum
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: 角色名称
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回角色列表
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 角色列表
      tags:
      - 角色管理
  /sysRole/updateDataScope:
    put:
      consumes:
      - application/json
      description: 更新角色的数据权限
      parameters:
      - description: 角色数据权限信息
        in: body
        name: roleDataScope
        required: true
        schema:
          $ref: '#/definitions/models.SysRoleDataScopeUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 角色数据权限更新成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 更新角色数据权限
      tags:
      - 角色管理
...
swagger: "2.0"